+++ lib/src/universal_ble.dart
@@ -9,17 +9,17 @@ import 'package:universal_ble/src/universal_ble_web/universal_ble_web.dart';
 import 'package:universal_ble/universal_ble.dart';
 
 class UniversalBle {
-  /// Get platform specific implementation
+  /// Get platform specific implementation.
   static UniversalBlePlatform _platform = _defaultPlatform();
   static final BleCommandQueue _bleCommandQueue = BleCommandQueue();
 
-  /// Set custom platform specific implementation (e.g. for testing)
+  /// Set custom platform specific implementation (e.g. for testing).
   static void setInstance(UniversalBlePlatform instance) =>
       _platform = instance;
 
   /// Set global timeout for all commands.
-  /// Default timeout is 10 seconds
-  /// Set to null to disable
+  /// Default timeout is 10 seconds.
+  /// Set to null to disable.
   static set timeout(Duration? duration) {
     _bleCommandQueue.timeout = duration;
   }
@@ -27,16 +27,16 @@ class UniversalBle {
   /// Set how commands will be executed. By default, all commands are executed in a global queue (`QueueType.global`),
   /// with each command waiting for the previous one to finish.
   ///
-  /// [QueueType.global] will execute commands of all devices in a single queue
-  /// [QueueType.perDevice] will execute command of each device in separate queues
-  /// [QueueType.none] will execute all commands in parallel
+  /// [QueueType.global] will execute commands of all devices in a single queue.
+  /// [QueueType.perDevice] will execute command of each device in separate queues.
+  /// [QueueType.none] will execute all commands in parallel.
   static set queueType(QueueType queueType) {
     _bleCommandQueue.queueType = queueType;
     UniversalBlePlatform.logInfo('Queue ${queueType.name}');
   }
 
-  /// Get Bluetooth availability state
-  /// To be notified of updates, set [onAvailabilityChange] listener
+  /// Get Bluetooth availability state.
+  /// To be notified of updates, set [onAvailabilityChange] listener.
   static Future<AvailabilityState> getBluetoothAvailabilityState() async {
     return await _bleCommandQueue.queueCommand(
       () => _platform.getBluetoothAvailabilityState(),
@@ -44,9 +44,9 @@ class UniversalBle {
   }
 
   /// Start scan.
-  /// Scan results will arrive in [onScanResult] listener
-  /// It might throw errors if Bluetooth is not available
-  /// `webRequestOptions` is supported on Web only
+  /// Scan results will arrive in [onScanResult] listener.
+  /// It might throw errors if Bluetooth is not available.
+  /// `webRequestOptions` is supported on Web only.
   static Future<void> startScan({
     ScanFilter? scanFilter,
     PlatformConfig? platformConfig,
@@ -61,8 +61,8 @@ class UniversalBle {
   }
 
   /// Stop scan.
-  /// Set [onScanResult] listener to `null` if you don't need it anymore
-  /// It might throw errors if Bluetooth is not available
+  /// Set [onScanResult] listener to `null` if you don't need it anymore.
+  /// It might throw errors if Bluetooth is not available.
   static Future<void> stopScan() async {
     return await _bleCommandQueue.queueCommand(
       () => _platform.stopScan(),
@@ -71,10 +71,10 @@ class UniversalBle {
   }
 
   /// Connect to a device.
-  /// Get notified of connection state changes in [onConnectionChange] listener
-  /// It is advised to stop scanning before connecting
-  /// It might throw errors if device is not connectable
-  /// `connectionTimeout` is supported on Web only
+  /// Get notified of connection state changes in [onConnectionChange] listener.
+  /// It is advised to stop scanning before connecting.
+  /// It might throw errors if device is not connectable.
+  /// `connectionTimeout` is supported on Web only.
   static Future<void> connect(
     String deviceId, {
     Duration? connectionTimeout,
@@ -86,7 +86,7 @@ class UniversalBle {
   }
 
   /// Disconnect from a device.
-  /// Get notified of connection state changes in [onConnectionChange] listener
+  /// Get notified of connection state changes in [onConnectionChange] listener.
   static Future<void> disconnect(String deviceId) async {
     return await _bleCommandQueue.queueCommand(
       () => _platform.disconnect(deviceId),
@@ -94,7 +94,7 @@ class UniversalBle {
     );
   }
 
-  /// Discover services of a device
+  /// Discover services of a device.
   static Future<List<BleService>> discoverServices(String deviceId) async {
     return await _bleCommandQueue.queueCommand(
       () => _platform.discoverServices(deviceId),
@@ -103,9 +103,9 @@ class UniversalBle {
   }
 
   /// Set a characteristic notifiable.
-  /// Set `bleInputProperty` to [BleInputProperty.notification] or [BleInputProperty.indication]
-  /// Updates will arrive in [onValueChange] listener
-  /// To stop listening to a characteristic, set `bleInputProperty` to [BleInputProperty.disabled]
+  /// Set `bleInputProperty` to [BleInputProperty.notification] or [BleInputProperty.indication].
+  /// Updates will arrive in [onValueChange] listener.
+  /// To stop listening to a characteristic, set `bleInputProperty` to [BleInputProperty.disabled].
   static Future<void> setNotifiable(
     String deviceId,
     String service,
@@ -123,8 +123,8 @@ class UniversalBle {
     );
   }
 
-  /// Read a characteristic value
-  /// On iOS and MacOS this command will also trigger [onValueChange] listener
+  /// Read a characteristic value.
+  /// On iOS and MacOS this command will also trigger [onValueChange] listener.
   static Future<Uint8List> readValue(
     String deviceId,
     String service,
@@ -140,8 +140,8 @@ class UniversalBle {
     );
   }
 
-  /// Write a characteristic value
-  /// To write a characteristic value with response, set `bleOutputProperty` to [BleOutputProperty.withResponse]
+  /// Write a characteristic value.
+  /// To write a characteristic value with response, set `bleOutputProperty` to [BleOutputProperty.withResponse].
   static Future<void> writeValue(
     String deviceId,
     String service,
@@ -161,8 +161,8 @@ class UniversalBle {
     );
   }
 
-  /// Request MTU value
-  /// `requestMtu` is not supported on `Linux` and `Web
+  /// Request MTU value.
+  /// `requestMtu` is not supported on `Linux` and `Web.
   static Future<int> requestMtu(String deviceId, int expectedMtu) async {
     return await _bleCommandQueue.queueCommand(
       () => _platform.requestMtu(deviceId, expectedMtu),
@@ -170,8 +170,8 @@ class UniversalBle {
     );
   }
 
-  /// Check if a device is paired
-  /// Returns null on `Apple` and `Web`
+  /// Check if a device is paired.
+  /// Returns null on `Apple` and `Web`.
   static Future<bool?> isPaired(String deviceId) async {
     if (kIsWeb || Platform.isIOS || Platform.isMacOS) return null;
     return await _bleCommandQueue.queueCommand(
@@ -191,8 +191,8 @@ class UniversalBle {
     );
   }
 
-  /// Unpair a device
-  /// It might throw an error if device is not paired
+  /// Unpair a device.
+  /// It might throw an error if device is not paired.
   static Future<void> unPair(String deviceId) async {
     return await _bleCommandQueue.queueCommand(
       () => _platform.unPair(deviceId),
@@ -200,11 +200,11 @@ class UniversalBle {
     );
   }
 
-  /// Get connected devices to the system (connected by any app)
-  /// Use [withServices] to filter devices by services
-  /// On `Apple`, [withServices] is required to get connected devices, else [1800] service will be used as default filter
-  /// On `Android`, `Linux` and `Windows`, if [withServices] is used, then internally all services will be discovered for each device first (either by connecting or by using cached services)
-  /// Not supported on `Web`
+  /// Get connected devices to the system (connected by any app).
+  /// Use [withServices] to filter devices by services.
+  /// On `Apple`, [withServices] is required to get connected devices, else [1800] service will be used as default filter.
+  /// On `Android`, `Linux` and `Windows`, if [withServices] is used, then internally all services will be discovered for each device first (either by connecting or by using cached services).
+  /// Not supported on `Web`.
   static Future<List<BleDevice>> getSystemDevices({
     List<String>? withServices,
   }) async {
@@ -222,9 +222,9 @@ class UniversalBle {
     );
   }
 
-  /// Enable Bluetooth
-  /// It might throw errors if Bluetooth is not available
-  /// Not supported on `Web` and `Apple`
+  /// Enable Bluetooth.
+  /// It might throw errors if Bluetooth is not available.
+  /// Not supported on `Web` and `Apple`.
   static Future<bool> enableBluetooth() async {
     return await _bleCommandQueue.queueCommand(
       () => _platform.enableBluetooth(),
@@ -241,7 +241,7 @@ class UniversalBle {
   static bool receivesAdvertisements(String deviceId) =>
       _platform.receivesAdvertisements(deviceId);
 
-  /// Get Bluetooth state availability
+  /// Get Bluetooth state availability.
   static set onAvailabilityChange(OnAvailabilityChange? onAvailabilityChange) {
     _platform.onAvailabilityChange = onAvailabilityChange;
     if (onAvailabilityChange != null) {
@@ -251,23 +251,23 @@ class UniversalBle {
     }
   }
 
-  /// Get updates of remaining items of a queue
+  /// Get updates of remaining items of a queue.
   static set onQueueUpdate(OnQueueUpdate? onQueueUpdate) =>
       _bleCommandQueue.onQueueUpdate = onQueueUpdate;
 
-  /// Get scan results
+  /// Get scan results.
   static set onScanResult(OnScanResult? bleDevice) =>
       _platform.onScanResult = bleDevice;
 
-  /// Get connection state changes
+  /// Get connection state changes.
   static set onConnectionChange(OnConnectionChange? onConnectionChange) =>
       _platform.onConnectionChange = onConnectionChange;
 
-  /// Get characteristic value updates, set `bleInputProperty` in [setNotifiable] to [BleInputProperty.notification] or [BleInputProperty.indication]
+  /// Get characteristic value updates, set `bleInputProperty` in [setNotifiable] to [BleInputProperty.notification] or [BleInputProperty.indication].
   static set onValueChange(OnValueChange? onValueChange) =>
       _platform.onValueChange = onValueChange;
 
-  /// Get pair state changes,
+  /// Get pair state changes.
   static set onPairingStateChange(OnPairingStateChange pairingStateChange) =>
       _platform.onPairingStateChange = pairingStateChange;